Either was used then 3 choices were presented
Original text:
you can install Azurite by using either Node Package Manager, DockerHub, or by cloning the Azurite GitHub repository

There is no need to mention the three ways to install/run. The tabs are right there so point the developer at the tabs. Also the above sentence does not understand the install process. npm and github clone are the same thing but they are the first and third items int he list. DockerHub is a containerized experience not a local install. It should not be sandwiched between the local install options.